fixed test shouldOpenCircuitBreakerAndCloseAfterSleepTime
Root cause:
WireMockClassRule starts service per class and doesn't refresh it.
All stubs are still exists, so if different tests use the same endpoint,
it causes to unexpected results. In this case one of the tests checks
the number of invoked stubs in test and fails, because server returns all stubs in the suite.

In contrast, WireMockRule reloads context before each test

Resolution summary:
change WireMockClassRule to WireMockRule